Launch Survo editor

Initializes Survo R and opens the editorial interface window. For more information, type HELP? in the editor and activate by ESC-key. See also Survo R web pages and Survo web pages for more information. Especially getting started tour, Survo books, Survo demos and Survo help system are likely to be useful for a newbie.

Usage

survo(sucro = "<empty>", config = "<empty>")

Arguments

sucro

Path for start sucro.

config

Path for config file (SURVOR.APU).